refactor hooks and components
