+++ ui/src/features/Dashboard/components/MainSection/components/index.ts
@@ -1,0 +1,3 @@
+export { ProjectDeletionDialog } from "./ProjectDeletionDialog";
+export { ProjectEditDialog } from "./ProjectEditDialog";
+export { ProjectCard } from "./ProjectCard";
